backend/pyproject.toml
[tool.poetry]
name = "schoolflow-backend"
version = "0.1.0"
description = "SchoolFlow backend - integration skeleton + fee module"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.22.0" }
sqlalchemy = "^2.0.18"
alembic = "^1.11.1"
pydantic = "^2.9.0"
python-dotenv = "^1.0.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
jinja2 = "^3.1.2"
pdfkit = "^1.0.0"
requests = "^2.31.0"
httpx = "^0.27.0"
apscheduler = "^3.10.1"
tenacity = "^8.2.2"
email-validator = "^1.3.1"
psycopg2-binary = "^2.9.10"
pydantic-settings = "^2.10.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.2"
pytest-cov = "^4.1.0"
ruff = "^0.5.0"
black = "^24.1.0"
isort = "^5.12.0"
mypy = "^1.9.0"

[tool.poetry.scripts]
start = "uvicorn app.main:app --reload --port 8000"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
